requirements.txt
# FastAPI and Web Framework
fastapi
uvicorn
python-multipart
python-dotenv

# WorkOS Authentication
workos

# Supabase and Database
supabase
psycopg2-binary
pgvector

# AI and LLM
google-generativeai
langchain
langchain-google-genai
langgraph
huggingface-hub

# Embeddings and Vector Operations
transformers
torch
numpy

# Security and Encryption
cryptography
passlib[bcrypt]
python-jose[cryptography]

# HTTP and Requests
httpx
aiohttp

# Data Processing
pydantic
pandas

# Logging and Monitoring
structlog

# Development and Testing
pytest
pytest-asyncio

# Additional dependencies for therapy and mental exercise agents
elevenlabs  # ElevenLabs Conversational AI SDK
websockets  # For real-time communication with agents
python-json-logger  # Enhanced JSON logging for therapy sessions

# Image processing for nutrition agent
Pillow==10.1.0

# LangGraph for workflow orchestration (if not already present)
langgraph==0.0.69

# Additional dependencies for nutrition agent
httpx==0.25.2  # For USDA API calls
